V2.13.09

+++ PixelFlightFight/Globals.h
@@ -67,7 +67,7 @@
 // 飞机工坊中的单元像素边长
 #define UNIT_EDGE 10
 
-using InsId = int;
+using InsId = long long;
 using PlaneId = long long;
 
 //实体类型
